Finished the responsive design for the login page and also the whole project
Ok that took only 15 minutes to make.

So, I guess thats all.

The project has come to an end. It was a really cool project where I had a lot of fun but also a lot frustrations xD
But in the end I really enjoyed making this project and I learned a lot regarding webdesign and web development, webservers, databases and how difficult it is to keep track of larger projects.

I have a ton of Ideas what to implement next in this project but not today. I started this project, including the minecraft server for over one year, it was on 29.01.2023 the project started. Now, on 20.04.2024 i feel like i have to move on to new projects and challenges.
Im sure that I will come back to this project one day and make it even better but for now I'm done with it.

The next thing I will look into is mobile app development with Swift. I already have a cool project in mind I'm going to make with it, so it will be worth it to move on.

It may sound weird, but while I'm writing this Im kinda sad that the project is done for now, to be honest.
Tomorrow I will try to dockerize the whole thing and then its over.

It is 23:58 o'clock now and I don't really know what to write anymore, so I'm stopping right here and finally commit the last commit for this project for now.